# Optimize RLE decoding (part 4)