--- a/internal/bitpack/bitpack.go
+++ b/internal/bitpack/bitpack.go
@@ -1,0 +1,7 @@
+// Package bitpack implements efficient bit packing and unpacking routines for
+// integers of various bit widths.
+package bitpack
+
+func byteCount(bitCount uint) uint {
+	return (bitCount + 7) / 8
+}