--- a/internal/bitpack/unpack_int32_purego.go
+++ b/internal/bitpack/unpack_int32_purego.go
@@ -2,7 +2,9 @@
 
 package bitpack
 
-import "github.com/segmentio/parquet-go/internal/unsafecast"
+import (
+	"github.com/segmentio/parquet-go/internal/unsafecast"
+)
 
 func unpackInt32(dst []int32, src []byte, bitWidth uint) {
 	bits := unsafecast.BytesToUint32(src)