--- a/internal/bitpack/unpack_int64_amd64.s
+++ b/internal/bitpack/unpack_int64_amd64.s
@@ -1,0 +1,69 @@
+//go:build !purego
+
+#include "funcdata.h"
+#include "textflag.h"
+
+// func unpackInt64Default(dst []int64, src []uint32, bitWidth uint)
+TEXT ·unpackInt64Default(SB), NOSPLIT, $0-56
+    MOVQ dst_base+0(FP), AX
+    MOVQ dst_len+8(FP), DX
+    MOVQ src_base+24(FP), BX
+    MOVQ bitWidth+48(FP), CX
+
+    MOVQ $1, R8 // bitMask = (1 << bitWidth) - 1
+    SHLQ CX, R8, R8
+    DECQ R8
+    MOVQ CX, R9 // bitWidth
+
+    XORQ DI, DI // bitOffset
+    XORQ SI, SI // index
+    XORQ R10, R10
+    XORQ R11, R11
+    XORQ R14, R14
+    JMP test
+loop:
+    MOVQ DI, R10
+    MOVQ DI, CX
+    SHRQ $5, R10      // i = bitOffset / 32
+    ANDQ $0b11111, CX // j = bitOffset % 32
+
+    MOVLQZX (BX)(R10*4), R11
+    MOVQ R8, R12  // d = bitMask
+    SHLQ CX, R12  // d = d << j
+    ANDQ R12, R11 // d = src[i] & d
+    SHRQ CX, R11  // d = d >> j
+
+    MOVQ CX, R13
+    ADDQ R9, R13
+    CMPQ R13, $32
+    JBE next // j+bitWidth <= 32 ?
+    MOVQ CX, R15 // j
+
+    MOVLQZX 4(BX)(R10*4), R14
+    MOVQ $32, CX
+    SUBQ R15, CX  // k = 32 - j
+    MOVQ R8, R12  // c = bitMask
+    SHRQ CX, R12  // c = c >> k
+    ANDQ R12, R14 // c = src[i+1] & c
+    SHLQ CX, R14  // c = c << k
+    ORQ R14, R11  // d = d | c
+
+    CMPQ R13, $64
+    JBE next
+
+    MOVLQZX 8(BX)(R10*4), R14
+    MOVQ $64, CX
+    SUBQ R15, CX  // k = 64 - j
+    MOVQ R8, R12  // c = bitMask
+    SHRQ CX, R12  // c = c >> k
+    ANDQ R12, R14 // c = src[i+2] & c
+    SHLQ CX, R14  // c = c << k
+    ORQ R14, R11  // d = d | c
+next:
+    MOVQ R11, (AX)(SI*8) // dst[n] = d
+    ADDQ R9, DI          // bitOffset += bitWidth
+    INCQ SI
+test:
+    CMPQ SI, DX
+    JNE loop
+    RET