… encoding/delta/binary_packed_decode_amd64.s
… encoding/delta/binary_packed_masks_amd64.s
… encoding/delta/binary_packed_test.go
… encoding/rle/rle_amd64.s
… internal/bitpack/masks_int32_amd64.s
… internal/bitpack/unpack_int32_amd64.s
… internal/bitpack/unpack_int64_purego.go
… internal/bitpack/unpack_test.go
… page.go