This PR is a follow up to https://github.com/segmentio/parquet-go/pull/230, it moves the unpacking routines from `encoding/delta` to `internal/bitpack` to reuse the functions in `encoding/rle` and optimize the decoding of dictionaries:

```
name                         old time/op    new time/op     delta
Decode/RLE/int32               25.4µs ± 1%      1.1µs ± 0%    -95.48%  (p=0.000 n=8+10)
Decode/RLE_DICTIONARY/int32    25.0µs ± 0%      1.1µs ± 0%    -95.41%  (p=0.000 n=10+10)

name                         old speed      new speed       delta
Decode/RLE/int32             1.57GB/s ± 1%  34.80GB/s ± 0%  +2110.13%  (p=0.000 n=8+10)
Decode/RLE_DICTIONARY/int32  1.60GB/s ± 0%  34.87GB/s ± 0%  +2078.80%  (p=0.000 n=10+10)

name                         old value/s    new value/s     delta
Decode/RLE/int32                 394M ± 1%      8702M ± 0%  +2110.13%  (p=0.000 n=8+10)
Decode/RLE_DICTIONARY/int32      400M ± 0%      8717M ± 0%  +2078.79%  (p=0.000 n=10+10)
```

Contributes to resolving https://github.com/segmentio/parquet-go/issues/226