@achille-roussel 2022-06-14
```
name                old time/op    new time/op     delta
Decode/RLE/boolean     105ns ± 1%      102ns ± 1%     -3.00%  (p=0.000 n=10+10)
Decode/RLE/levels     17.8µs ± 0%      0.7µs ± 0%    -96.13%  (p=0.000 n=9+9)
Decode/RLE/int32      25.3µs ± 0%      1.2µs ± 0%    -95.23%  (p=0.000 n=9+9)

name                old speed      new speed       delta
Decode/RLE/boolean  11.9GB/s ± 1%   12.2GB/s ± 1%     +3.11%  (p=0.000 n=10+10)
Decode/RLE/levels    563MB/s ± 0%  14535MB/s ± 0%  +2481.21%  (p=0.000 n=9+9)
Decode/RLE/int32    1.58GB/s ± 0%  33.11GB/s ± 1%  +1997.38%  (p=0.000 n=9+9)

name                old value/s    new value/s     delta
Decode/RLE/boolean     94.8G ± 1%      97.8G ± 1%     +3.11%  (p=0.000 n=10+10)
Decode/RLE/levels       563M ± 0%     14537M ± 0%  +2481.21%  (p=0.000 n=9+9)
Decode/RLE/int32        395M ± 0%      8278M ± 1%  +1997.38%  (p=0.000 n=9+9)
```